pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "optimiser_engine"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "PyYAML",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["optimiser_engine*"]
